Factor TestCase shared code
