+++ tests/test_npm.py
@@ -5,19 +5,12 @@
 
 """Test NPM-style specifications."""
 
-import unittest
-import sys
-
 from semantic_version import base
 
+from . import testing
 
-class NpmSpecTests(unittest.TestCase):
-    if sys.version_info[0] <= 2:
-        import contextlib
 
-        @contextlib.contextmanager
-        def subTest(self, **kwargs):
-            yield
+class NpmSpecTests(testing.TestCase):
 
     examples = {
         # range: [matchings], [failings]
